+++ lib/src/main/java/com/smileidentity/viewmodel/document/DocumentCaptureViewModel.kt
@@ -15,6 +15,7 @@ import com.google.mlkit.vision.objects.ObjectDetection
 import com.google.mlkit.vision.objects.ObjectDetector
 import com.google.mlkit.vision.objects.defaults.ObjectDetectorOptions
 import com.smileidentity.R
+import com.smileidentity.SmileIDCrashReporting
 import com.smileidentity.compose.document.DocumentCaptureSide
 import com.smileidentity.models.v2.DocumentImageOriginValue
 import com.smileidentity.models.v2.Metadatum
@@ -23,6 +24,8 @@ import com.smileidentity.util.createDocumentFile
 import com.smileidentity.util.postProcessImage
 import com.ujizin.camposer.state.CameraState
 import com.ujizin.camposer.state.ImageCaptureResult
+import io.sentry.Breadcrumb
+import io.sentry.SentryLevel
 import java.io.File
 import java.io.IOException
 import kotlin.math.abs
@@ -166,12 +169,30 @@ class DocumentCaptureViewModel(
                                 )
                             }
                         } catch (e: IOException) {
-                            Timber.e(e, "Error processing captured image")
+                            Timber.e(e, "IOException processing captured image")
+                            SmileIDCrashReporting.hub.captureException(e) {
+                                it.level = SentryLevel.INFO
+                                it.addBreadcrumb(
+                                    Breadcrumb(
+                                        "Smile ID DocumentCaptureViewModel " +
+                                            "IOException",
+                                    ),
+                                )
+                            }
                             _uiState.update {
                                 it.copy(captureError = e, showCaptureInProgress = false)
                             }
                         } catch (e: OutOfMemoryError) {
-                            Timber.e(e, "Out of memory while processing captured image")
+                            Timber.e(e, "OutOfMemoryError processing captured image")
+                            SmileIDCrashReporting.hub.captureException(e) {
+                                it.level = SentryLevel.INFO
+                                it.addBreadcrumb(
+                                    Breadcrumb(
+                                        "Smile ID DocumentCaptureViewModel " +
+                                            "OutOfMemoryError",
+                                    ),
+                                )
+                            }
                             _uiState.update {
                                 it.copy(captureError = e, showCaptureInProgress = false)
                             }
@@ -180,7 +201,16 @@ class DocumentCaptureViewModel(
                 }
 
                 is ImageCaptureResult.Error -> {
-                    Timber.e("Error capturing document", result.throwable)
+                    Timber.e("ImageCaptureResult.Error capturing document", result.throwable)
+                    SmileIDCrashReporting.hub.captureException(result.throwable) {
+                        it.level = SentryLevel.INFO
+                        it.addBreadcrumb(
+                            Breadcrumb(
+                                "Smile ID DocumentCaptureViewModel " +
+                                    "ImageCaptureResult.Error",
+                            ),
+                        )
+                    }
                     _uiState.update {
                         it.copy(captureError = result.throwable, showCaptureInProgress = false)
                     }